lazy load codeEditor component to reduce the frontend bundle size
