+++ workspaces/extensions/plugins/catalog-backend-module-extensions/report.api.md
@@ -11,6 +11,7 @@ import type { CatalogApi } from '@backstage/catalog-client';
 import { CatalogProcessor } from '@backstage/plugin-catalog-node';
 import { CatalogProcessorCache } from '@backstage/plugin-catalog-node';
 import { CatalogProcessorEmit } from '@backstage/plugin-catalog-node';
+import { Config } from '@backstage/config';
 import { DynamicPluginProvider } from '@backstage/backend-dynamic-feature-service';
 import { Entity } from '@backstage/catalog-model';
 import { EntityProvider } from '@backstage/plugin-catalog-node';
@@ -25,7 +26,7 @@ import { SchedulerServiceTaskRunner } from '@backstage/backend-plugin-api';
 
 // @public (undocumented)
 export abstract class BaseEntityProvider<T extends Entity> implements EntityProvider {
-    constructor(taskRunner: SchedulerServiceTaskRunner);
+    constructor(taskRunner: SchedulerServiceTaskRunner, config?: Config);
     // (undocumented)
     connect(connection: EntityProviderConnection): Promise<void>;
     // (undocumented)
